backend/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==2.3.3
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
psycopg2-binary
numpy==2.3.2
pandas==2.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.16.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
xgboost==3.0.5
gunicorn==21.2.0
sqlalchemy==2.0.21